Use dependency_links to install prompt_toolkit 2.0
To avoid having to install using a bash script, we can specify a version
of prompt_toolkit and provide the github branch tarball for 2.0 as a
dependency link.

This means when installing with pip install sml-sync
--process-dependency-links, pip will install prompt toolkit from the
github repo.

This means we can close the bitbucket repo.

Tested with pip 8.1 (gives a deprecation warning, but still works).